refactor: simplify translation system with unified common/locales/index.ts
- Create unified translation helper in common/locales/index.ts
- Remove frontend/locales/useTranslation.ts
- Remove frontend/locales/index.ts
- Remove common/locales/translateServer.ts
- Update all 16 frontend components/pages to import from common/locales
- Update all 5 backend managers to import from common/locales
- Single entry point for translation functions (useTranslation and t)
- Simplify architecture while maintaining type safety
- Reduce files from 3 utilities to 1 unified file
- All checks pass and builds succeed
