+++ backend/src/managers/timeEntry.manager.ts
@@ -1,5 +1,5 @@
 import type { Socket } from 'socket.io'
-import { t } from '../../../common/locales/translateServer'
+import { t } from '../../../common/locales'
 import { isPostLapTimeRequest } from '../../../common/models/requests'
 import type {
   BackendResponse,
