+++ frontend/app/components/SearchableDropdown.tsx
@@ -7,7 +7,7 @@ import {
   type SetStateAction,
 } from 'react'
 import { twMerge } from 'tailwind-merge'
-import { useTranslation } from '../../locales/useTranslation'
+import { useTranslation } from '../../../common/locales'
 import { Button } from './Button'
 
 export type LookupItem = {
