+++ frontend/app/pages/Login.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState, type FormEvent } from 'react'
 import { useNavigate, useSearchParams } from 'react-router-dom'
+import { useTranslation } from '../../../common/locales'
 import { useAuth } from '../../contexts/AuthContext'
-import { useTranslation } from '../../locales/useTranslation'
 import { Button } from '../components/Button'
 import UserForm from '../components/UserForm'
