+++ frontend/app/pages/Track.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
+import { useTranslation } from '../../../common/locales'
 import type { GetLeaderboardRequest } from '../../../common/models/requests'
 import {
   type ErrorResponse,
@@ -11,7 +12,6 @@ import { WS_GET_LEADERBOARD } from '../../../common/utils/constants'
 import { formatTrackName } from '../../../common/utils/track'
 import { useAuth } from '../../contexts/AuthContext'
 import { useConnection } from '../../contexts/ConnectionContext'
-import { useTranslation } from '../../locales/useTranslation'
 import LeaderboardView from '../components/Leaderboard'
 import LoadingView from '../components/LoadingView'
 import TrackTag from '../components/TrackTag'
