--- frontend/locales/index.ts
@@ -1,1 +1,0 @@
-export { useTranslation } from './useTranslation'
